move release models to separate files
